P4ADEV-1981 moved class
