+++ src/main/java/it/gov/pagopa/payhub/activities/connector/debtposition/TransferService.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.payhub.activities.connector.transfer;
+package it.gov.pagopa.payhub.activities.connector.debtposition;
 
 import it.gov.pagopa.payhub.activities.dto.classifications.TransferSemanticKeyDTO;
 import it.gov.pagopa.pu.debtposition.dto.generated.DebtPositionDTO;
