+++ src/main/java/it/gov/pagopa/payhub/activities/connector/debtposition/config/DebtPositionApisHolder.java
@@ -2,6 +2,8 @@ package it.gov.pagopa.payhub.activities.connector.debtposition.config;
 
 import it.gov.pagopa.pu.debtposition.client.generated.DebtPositionApi;
 import it.gov.pagopa.pu.debtposition.client.generated.DebtPositionSearchControllerApi;
+import it.gov.pagopa.pu.debtposition.client.generated.TransferApi;
+import it.gov.pagopa.pu.debtposition.client.generated.TransferSearchControllerApi;
 import it.gov.pagopa.pu.debtposition.generated.ApiClient;
 import it.gov.pagopa.pu.debtposition.generated.BaseApi;
 import jakarta.annotation.PreDestroy;
@@ -14,23 +16,24 @@ import org.springframework.web.client.RestTemplate;
 @Lazy
 @Service
 public class DebtPositionApisHolder {
-
+    private final TransferSearchControllerApi transferSearchControllerApi;
+    private final TransferApi transferApi;
     private final DebtPositionSearchControllerApi debtPositionSearchControllerApi;
-
     private final DebtPositionApi debtPositionApi;
-
     private final ThreadLocal<String> bearerTokenHolder = new ThreadLocal<>();
 
     public DebtPositionApisHolder(
-            @Value("${rest.debt-position.base-url}") String baseUrl,
-            RestTemplateBuilder restTemplateBuilder) {
-        RestTemplate restTemplate = restTemplateBuilder.build();
+	    @Value("${rest.debt-position.base-url}") String baseUrl,
+	    RestTemplateBuilder restTemplateBuilder) {
+	    RestTemplate restTemplate = restTemplateBuilder.build();
         ApiClient apiClient = new ApiClient(restTemplate);
         apiClient.setBasePath(baseUrl);
         apiClient.setBearerToken(bearerTokenHolder::get);
 
         this.debtPositionSearchControllerApi = new DebtPositionSearchControllerApi(apiClient);
         this.debtPositionApi = new DebtPositionApi(apiClient);
+        this.transferSearchControllerApi = new TransferSearchControllerApi(apiClient);
+        this.transferApi = new TransferApi(apiClient);
     }
 
     @PreDestroy
@@ -47,6 +50,17 @@ public class DebtPositionApisHolder {
 
     public DebtPositionApi getDebtPositionApi(String accessToken){
         return getApi(accessToken, debtPositionApi);
+    }
+
+    /**
+     * It will return a {@link TransferSearchControllerApi} instrumented with the provided accessToken. Use null if auth is not required
+     */
+    public TransferSearchControllerApi getTransferSearchControllerApi(String accessToken) {
+        return getApi(accessToken, transferSearchControllerApi);
+    }
+
+    public TransferApi getTransferApi(String accessToken){
+        return getApi(accessToken, transferApi);
     }
 
     private <T extends BaseApi> T getApi(String accessToken, T api) {
